Removing src/rgl/hash.* since these are duplicates of src/Game/Hash.*.
This unfortunately means that src/rgl now has a dependency on src/Game
which it didn't before. It may be a circular dependency too and I hate
those. I'm prepared to live with it for the moment though because
"rgl is going away".

Sorry to the Linux guys: I'm 99% sure I've hosed your build. "All" you
need to do is change the src/rgl Makefile configuration to pull in
#includes from src/Game, specifically src/Game/Hash.h.


git-svn-id: svn://www.homeworldsdl.org:3692/homeworldsdl/homeworld/trunk@311 026c9d8a-83c9-0310-a9c7-971d0a006279