Finally commit some casts relating to gcc v4 changes.
Large number of changes to SaveGame.c all of which are casts to stop the
large number of warnings. This seems to work, but any odd savegame
strangeness let me know.

This will still fail to quit under Linux due to 
http://homesource.nekomimicon.net/sourceforum/viewtopic.php?t=150

added mission05_OEM to Makefile.am Has no effect on the game but it's
there if we (eventually) want it. :)




git-svn-id: svn://www.homeworldsdl.org:3692/homeworldsdl/homeworld/trunk@364 026c9d8a-83c9-0310-a9c7-971d0a006279