Rewrite of the cameraRayCast() function from first principles in order to fix mouse cursor alignment with the pieplate movement disc when rendering widescreen aspect ratios.
Unfortunately alignment still isn't quite right (for any aspect ratio)
as there's an annoying fudge factor in there to get things looking right.
I suspect it's an issue with the orthogonal reference frames the function
is dealing with vs the perspective projection of the game environment.
Tracking it down will take a fair bit of effort though and the code here
is a marked improvement over what we currently render so I'm committing
it now. 


git-svn-id: svn://www.homeworldsdl.org:3692/homeworldsdl/homeworld/trunk@475 026c9d8a-83c9-0310-a9c7-971d0a006279