Filename pattern support in bigExtract.
I have assumed that all platforms except Windows provide the POSIX function
fnmatch.  Windows is currently not supported: there all files are extracted
as before.  If you know what Windows uses instead of fnmatch and are
able to test there, please fix this.

Filelists (@-notation) are not supported (as they are in bigAdd).
Patterns are matched against the entire pathname, not just the basename.
The directory separator is '\', which is also the escape character in
patterns, resulting in some leaning toothpick syndrome.
Matching is case-sensitive.


git-svn-id: svn://www.homeworldsdl.org:3692/homeworldsdl/homeworld/trunk@608 026c9d8a-83c9-0310-a9c7-971d0a006279