__BEGIN_X86_64_COMMIT__
Initial commit of all changes to the header files.
Additional changes to the sound code are all in the
definitions of the variable types. Use of long and
unsigned long are different sizes within the different
architectures. Incuded Types.h to standardise on
sdword and udword instead.

I've adopted a type of memsize and smemsize which will
be difined on the size of the pointers. for x86 it is 
udword and sdwords, but for x86_64 it'll be uqword and 
sqword. 
These are the variables which may need to hold a pointer.

Tested and all seems well, but cannot test everything.

Updates to follow:
1) Changes to the configure and makefile templates
2) updates to the code.(staggered by chunks).
3) Stuff I've forgotten to make it work....

Aunxx.




git-svn-id: svn://www.homeworldsdl.org:3692/homeworldsdl/homeworld/trunk@659 026c9d8a-83c9-0310-a9c7-971d0a006279