Changes to configure and Makefile templates.
Tested for x86 and seems okay. will test for x86_64
after commit.

Next commit will be the start of the .c files but there
is a non-trivial ammount of changes to include.......

Probably be about 2 days for the next commit.
Please test and let me know what/if anything seems broken.

Aunxx.




git-svn-id: svn://www.homeworldsdl.org:3692/homeworldsdl/homeworld/trunk@660 026c9d8a-83c9-0310-a9c7-971d0a006279