Turning off optimisation for Mac OS X debugging builds
Previously reverted in r676 thanks to Xcode 3's "2.4 compatible" mode.

Xcode "Optimization Level" description:
None: Do not optimize. [-O0]

With this setting, the compiler's goal is to reduce the cost of compilation
and to make debugging produce the expected results. Statements are
independent: if you stop the program with a breakpoint between statements,
you can then assign a new value to any variable or change the program
counter to any other statement in the function and get exactly the results
you would expect from the source code.

git-svn-id: svn://www.homeworldsdl.org:3692/homeworldsdl/homeworld/trunk@718 026c9d8a-83c9-0310-a9c7-971d0a006279