fix: Build with --enable-sanitizers
With sanitizers enabled, the game now gets to the main menu and is able
to load non-trivial saves without crashing.
I'm hoping having builds with sanitizers enabled will help debugging as
you don't have to run the game through gdb to get a backtrace in case of
crashes. So this is accessible to less tech-savvy people.

fix: linux: Compiling with clang

nix: fix: dlopen() with sanitizers

---

I rewrote feFindChildren mostly to understand it, but the problem was
getting the `atom` from region->userID, and not region->atom