Added UV interpolation in morphed mesh rendering
The meshMorphedObjectRender() takes two frames of an animation (based on different meshes and textures) interpolates between them and renders them. This requires a certain set of requirements to be met such as the two meshes having the same amount of vertices+polygons and the vertex indices of the polygons matching. The code also mandates that the UV coordinates at the vertices match and leverages that by just using the UV coordinates of the base mesh (first frame of animation).

Turns out the ion beam (bulletionray.ebg) has 4 frames defined:

```
label(beamMesh)
   morphAnimDWORD(ETG\meshes\misc\beamRibbon.geo, 0)
   morphAnimDWORD(ETG\meshes\misc\beamRibbon01.geo, 1)
   morphAnimDWORD(ETG\meshes\misc\beamRibbon01.geo, 0)
   morphAnimDWORD(ETG\meshes\misc\beamRibbon01.geo, 2)
```

The 0,1,2 at the end define flags. 2 meaning the end of a loop, 1 beginning of a loop. Effectively the animation only iterates through frames 1,2,3 and never uses frame 0. The ebg file also sets the start frame to 1 skipping 0 completely. But frame 0 is still loaded as the base mesh.

Now it turns out that beamRibbon.geo and beamRibbon01.geo do not have matching UVs - and since the rendering code relies on the UVs of the base mesh we have the weird rendering of tip of the ion beam were the UVs diverge. It looks like the ion beam animation was changed at one point to start at frame 1, probably because it looked very weird with the tip of the beam flickering due to the UV mismatch.

Now one could either fix this by correcting the content. Either through modified .ebg files (there are multiple beams with the same issue). By editing the geometry files to match the UVs - or by making the mesh morph rendering code use the actual UVs from the animated meshes and not use the static base mesh just for them.

This is what this MR does: UVs are now interpolated between animation frames in the same way as vertices and normals are. This fixes the rendering and makes the system a little bit more correct and powerful.

In principle this should not cause any new issues (hopefully there are no completely broken particle meshes).