Restore original rgl shading code to fix gas/nebulae rendering and lighting
These changes restore the original rgl vertex shading code. Instead of relying on opengl material and lighting code to generate vertex colors the internal CPU code path is used to calculate the vertex colors. This restores the original brightness of the shadowed side of meshes and the specular highlight on the nebulae. Probably also other things.

This has the unfortunate side effect of removing someone's previous attempt of modernizing the render path using VBOs (vertex buffers on the GPU). When we get to implement a completely alternate modern opengl render path that reimplements the CPU side vertex shading as GPU vertex and pixel shader we will return to gpu buffers but for the sake of correctness they have to go for now.