Fix key lookup when entering text
SDL keycodes correspond to the character that should be generated when
a key is pressed. It is unnecessary to convert this into a readable
key name to determine which key has been pressed. More than this, in
some cases it is incorrect -- the readable name for the space key, for
example, is "Space", and using the first character of this string to
determine what character to insert results in an 's', not a space!
This error would also cause function keys to insert an 'f', rather
than being ignored.

This commit also ignores any characters whose byte representation lies
outside the range from 0 to KEY_TOAL_KEYS, as we do not have a
character mapping for such keys.

Fixes #63.